# Reduce allocations and internal copies