--- a/src/PolygonClipper/Segment.cs
+++ b/src/PolygonClipper/Segment.cs
@@ -17,7 +17,7 @@ internal readonly struct Segment : IEquatable<Segment>
     /// <param name="source">The segment source.</param>
     /// <param name="target">The segment target.</param>
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public Segment(Vertex source, Vertex target)
+    public Segment(in Vertex source, in Vertex target)
     {
         this.Source = source;
         this.Target = target;
@@ -46,11 +46,11 @@ internal readonly struct Segment : IEquatable<Segment>
     public Vertex Max { get; }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public static bool operator ==(Segment left, Segment right)
+    public static bool operator ==(in Segment left, in Segment right)
         => left.Equals(right);
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public static bool operator !=(Segment left, Segment right)
+    public static bool operator !=(in Segment left, in Segment right)
         => !(left == right);
 
     /// <summary>