… src/PolygonClipper/Vertex.cs
… tests/PolygonClipper.Benchmarks/ClippingLibraryComparison.cs
… tests/PolygonClipper.Tests/TestPolygonUtilities.cs
… tests/PolygonClipper.Tests/TestPolygonUtilitiesTests.cs
… tests/TestData/Benchmarks/asia.geojson
… tests/TestData/Benchmarks/states_source.geojson