Some more optimization work @stefannikolei 

* Precomputed list capacities to reduce dynamic resizing and allocations.
* Applied the `in` modifier for `readonly struct` parameters to avoid unnecessary copies.
* Switched from a binary to a 4-ary heap for the priority queue, matching .NET’s `PriorityQueue` design for improved cache locality and shallower tree depth.
* Changed sweep event processing to accept unordered input, performing a single-pass heap construction (heapify) just before the sweep phase.

I also removed Clipper2 from the benchmarks as I was unable to get reliable output from the geometry. 

### Benchmarks.

```
BenchmarkDotNet v0.14.0, Windows 11 (10.0.26100.4652)
11th Gen Intel Core i7-11370H 3.30GHz, 1 CPU, 8 logical and 4 physical cores
.NET SDK 9.0.302
  [Host]     : .NET 9.0.7 (9.0.725.31616), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI
  DefaultJob : .NET 9.0.7 (9.0.725.31616), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI
```

Main

| Method         | Mean     | Error    | StdDev   | Ratio | RatioSD | Gen0   | Gen1   | Allocated |
|--------------- |---------:|---------:|---------:|------:|--------:|-------:|-------:|----------:|
| PolygonClipper | 44.23 us | 0.246 us | 0.218 us |  2.60 |    0.02 | 6.7139 | 0.4883 |  41.34 KB |

PR

| Method         | Mean     | Error    | StdDev   | Ratio | RatioSD | Gen0   | Gen1   | Allocated |
|--------------- |---------:|---------:|---------:|------:|--------:|-------:|-------:|----------:|
| PolygonClipper | 41.92 us | 0.178 us | 0.139 us |  2.35 |    0.03 | 5.7983 | 0.3052 |   35.8 KB | 


#### Comparison to other language implementations


| Method         | File                 | Mean         | Error      | StdDev     | Op/s      | Gen0      | Gen1     | Gen2     | Allocated  |
|--------------- |--------------------- |-------------:|-----------:|-----------:|----------:|----------:|---------:|---------:|-----------:|
| PolygonClipper | asia.geojson         | 29,136.01 us | 280.796 us | 234.477 us |     34.32 | 1156.2500 | 937.5000 | 468.7500 | 7277.38 KB |
| PolygonClipper | hole_hole.geojson    |     12.49 us |   0.130 us |   0.122 us | 80,061.19 |    1.7395 |   0.0305 |        - |    10.7 KB |
| PolygonClipper | state(...)ojson [21] |  1,485.58 us |   6.812 us |   5.318 us |    673.14 |   95.7031 |  66.4063 |        - |  597.38 KB |

[rust-geo-booleanop](https://github.com/21re/rust-geo-booleanop)

We're around .05x - 2.3x faster than the Rust implementation.
```
benches/hole_hole/union time:   [29.927 µs 32.195 µs 34.983 µs]
benches/state_source/union  time:   [2.6862 ms 3.0259 ms 3.3598 ms]
benches/asia/union      time:   [30.689 ms 33.651 ms 36.414 ms]
```

[Martinez](https://github.com/w8r/martinez?tab=readme-ov-file#benchmarks)

We're around 2.7x - 3.8x faster than the JavaScript implementation.
```
Hole_Hole
Martinez x 29,530 ops/sec ±1.65% (85 runs sampled)
JSTS x 2,051 ops/sec ±2.62% (85 runs sampled)
- Fastest is Martinez

Asia union
Martinez x 9.19 ops/sec ±3.30% (26 runs sampled)
JSTS x 7.60 ops/sec ±4.24% (23 runs sampled)
- Fastest is Martinez

States clip
Martinez x 227 ops/sec ±1.10% (82 runs sampled)
JSTS x 100 ops/sec ±2.54% (73 runs sampled)
- Fastest is Martinez
```